refactor: relocate ClaudeRunner to src/agent/runners/claude/runner.ts
Mirrors the PiRunner layout established in #644. Pure file move +
2-line import update — no behavioral change.

Part of #632 segment 2 (kernel relocation).
